Extract BindingIndex helper and use FakeToNullModule
- Extract parseIndex to shared BindingIndex::parse()
- Create FakeToNullModule instead of anonymous class
- Tighten assertion to check tree format
