+++ tests/di/Fake/FakeToNullModule.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ray\Di;
+
+class FakeToNullModule extends AbstractModule
+{
+    protected function configure(): void
+    {
+        $this->bind(FakeRobotInterface::class)->toNull();
+    }
+}
