+++ tests/di/ModuleJsonTest.php
@@ -71,12 +71,7 @@ class ModuleJsonTest extends TestCase
 
     public function testToNullBinding(): void
     {
-        $module = new class extends AbstractModule {
-            protected function configure(): void
-            {
-                $this->bind(FakeRobotInterface::class)->toNull();
-            }
-        };
+        $module = new FakeToNullModule();
         $json = $module->toJson();
         /** @var array{bindings: list<array{interface: string, name: string, type: string, to: mixed}>} $decoded */
         $decoded = json_decode($json, true);
